release 4.14-beta1

+++ demo/L1_Voronoi_diagram_2/include/CGAL/L1_voronoi_traits_2.h
@@ -12,8 +12,8 @@
 // This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
 // WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 //
-// $URL$
-// $Id$
+// $URL: https://github.com/CGAL/cgal/blob/releases/CGAL-4.14-beta1/GraphicsView/demo/L1_Voronoi_diagram_2/include/CGAL/L1_voronoi_traits_2.h $
+// $Id: L1_voronoi_traits_2.h 962ff6e %aI Marc Glisse
 // SPDX-License-Identifier: GPL-3.0+
 //
 // Author(s)     : Ophir Setter
@@ -85,8 +85,8 @@ public:
     FT delta_x = pp2->x() - pp1->x();
     FT delta_y = pp2->y() - pp1->y();
     
-    FT sign_x = CGAL::sign(delta_x);
-    FT sign_y = CGAL::sign(delta_y);
+    FT sign_x = (int)CGAL::sign(delta_x);
+    FT sign_y = (int)CGAL::sign(delta_y);
 
     FT abs_x = CGAL::abs(delta_x);
     FT abs_y = CGAL::abs(delta_y);
