+++ include/CGAL/Kernel_d/HyperplaneHd_impl.h
@@ -16,8 +16,8 @@
 // This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
 // WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 //
-// $URL$
-// $Id$
+// $URL: https://github.com/CGAL/cgal/blob/releases/CGAL-4.14-beta1/Kernel_d/include/CGAL/Kernel_d/HyperplaneHd_impl.h $
+// $Id: HyperplaneHd_impl.h 41f420f %aI Marc Glisse
 // SPDX-License-Identifier: LGPL-3.0+
 // 
 //
@@ -85,8 +85,8 @@ strong_cmp(const HyperplaneHd<RT,LA>& h1,
   int c1 = CGAL_NTS sign(h1.coefficient(i));
   int c2 = CGAL_NTS sign(h2.coefficient(i));
   if (c1 != c2) return CGAL_NTS compare(c1,c2);
-  RT s1 = (RT)CGAL_NTS sign(h2.coefficient(i)) * h2.coefficient(i); 
-  RT s2 = (RT)CGAL_NTS sign(h1.coefficient(i)) * h1.coefficient(i);
+  RT s1 = (RT)(int)CGAL_NTS sign(h2.coefficient(i)) * h2.coefficient(i);
+  RT s2 = (RT)(int)CGAL_NTS sign(h1.coefficient(i)) * h1.coefficient(i);
 
   Comparison_result c;
   while (++i <= d) {
